requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
bcrypt==4.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
coreapi==2.3.3
coreschema==0.0.4
decouple==0.0.7
distro==1.9.0
Django==5.2.1
django-cors-headers==4.7.0
djangorestframework==3.16.0
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.5.1
groq==0.26.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.6
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
openapi-codec==1.3.2
psycopg2-binary==2.9.10
pydantic==2.11.5
pydantic_core==2.33.2
PyJWT==2.10.1
python-decouple==3.8
python-dotenv==1.1.0
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.25.1
setuptools==80.9.0
simplejson==3.20.1
sniffio==1.3.1
sqlparse==0.5.3
typing-inspection==0.4.1
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
